requirements.txt
pdfplumber
streamlit
google-generativeai
python-dotenv
